refactor(coaches): migrate lineups to AppPageLayout
Extract useCoachLineupsPageLogic hook and CoachLineupsContainer.
Switches season source to useCoachCategory() selectedSeason, removing
the local useFetchSeasons + useEffect pattern.
